Move mutability error reporting to its own file
